wip working dynamic args
